requirements.txt
# Edge News Summarizer - Requirements
# Python 3.11+, CUDA 12.8+
# Generated from conda environment: my_sft

# Core ML Framework
torch==2.7.1+cu128
torchvision==0.22.1+cu128
torchaudio==2.7.1+cu128

# Transformers & Fine-tuning
transformers==5.2.0
peft==0.18.1
accelerate==1.11.0
bitsandbytes==0.49.2
autoawq

# Data Processing
datasets==4.0.0
pandas==2.3.3
numpy==2.3.5

# NLP & Evaluation
rouge-score==0.1.2
rouge-chinese==1.0.3
jieba==0.42.1
nltk==3.9.3

# Utilities
PyYAML==6.0.3
scipy==1.17.1
requests==2.32.5
python-dotenv==1.2.1
tqdm==4.67.3

# Optional: API clients (if using external services)
openai==2.24.0
huggingface_hub==1.5.0

# Development tools (optional)
# pytest>=7.4.0
# black>=23.0.0
# ruff==0.15.4

# Note: LLaMA-Factory should be installed separately:
# git clone https://github.com/hiyouga/LLaMA-Factory.git
# cd LLaMA-Factory && pip install -e .
